# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KhataPro CRM API — digital ledger platform for small businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products for a business
      parameters:
        - name: business_id
          in: query
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: low_stock
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/barcode/{barcode}:
    get:
      operationId: getProductByBarcode
      tags: [products]
      summary: Look up a product by barcode (for billing scan)
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  # ─── VENDORS ─────────────────────────────────────────────────────────────────
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List purchases for a business
      parameters:
        - name: business_id
          in: query
          required: true
          schema:
            type: integer
        - name: vendor_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [paid, pending, partial]
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseListResponse"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Record a purchase from a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Purchase created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get a purchase by ID, including line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "404":
          description: Not found
    put:
      operationId: updatePurchase
      tags: [purchases]
      summary: Update purchase payment/status/meta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseUpdate"
      responses:
        "200":
          description: Updated purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
    delete:
      operationId: deletePurchase
      tags: [purchases]
      summary: Delete a purchase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  # ─── SALES ORDERS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    Product:
      type: object
      required: [id, business_id, name, unit, gst_rate, cost_price, selling_price, stock_qty, created_at]
      properties:
        id:
          type: integer
        business_id:
          type: integer
        name:
          type: string
        barcode:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        unit:
          type: string
          enum: [pcs, kg, g, l, ml, pkt, box, bottle, dozen]
        hsn_code:
          type: ["string", "null"]
        gst_rate:
          type: number
        cost_price:
          type: number
        selling_price:
          type: number
        stock_qty:
          type: integer
        low_stock_alert:
          type: integer
        image:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [business_id, name, unit, selling_price]
      properties:
        business_id:
          type: integer
        name:
          type: string
        barcode:
          type: string
        sku:
          type: string
        category:
          type: string
        brand:
          type: string
        unit:
          type: string
          enum: [pcs, kg, g, l, ml, pkt, box, bottle, dozen]
          default: pcs
        hsn_code:
          type: string
        gst_rate:
          type: number
          default: 0
        cost_price:
          type: number
          default: 0
        selling_price:
          type: number
        stock_qty:
          type: integer
          default: 0
        low_stock_alert:
          type: integer
          default: 5
        image:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        barcode:
          type: string
        sku:
          type: string
        category:
          type: string
        brand:
          type: string
        unit:
          type: string
          enum: [pcs, kg, g, l, ml, pkt, box, bottle, dozen]
        hsn_code:
          type: string
        gst_rate:
          type: number
        cost_price:
          type: number
        selling_price:
          type: number
        stock_qty:
          type: integer
        low_stock_alert:
          type: integer
        image:
          type: string
    ProductListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Vendor ────────────────────────────────────────────────────────────────
...
    Purchase:
      type: object
      required: [id, business_id, vendor_id, amount, tax, amount_paid, status, entry_date, created_at]
      properties:
        id:
          type: integer
        business_id:
          type: integer
        vendor_id:
          type: integer
        vendor_name:
          type: string
        product_names:
          type: array
          items:
            type: string
        invoice_no:
          type: ["string", "null"]
        amount:
          type: number
        tax:
          type: number
        amount_paid:
          type: number
        status:
          type: string
          enum: [paid, pending, partial]
        bill_image_url:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        entry_date:
          type: string
          format: date
        product_count:
          type: integer
        created_by:
          type: integer
        created_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
    PurchaseInput:
      type: object
      required: [business_id, vendor_id, items]
      properties:
        business_id:
          type: integer
        vendor_id:
          type: integer
        invoice_no:
          type: string
        tax:
          type: number
          default: 0
        amount_paid:
          type: number
          default: 0
        bill_image_url:
          type: string
        description:
          type: string
        entry_date:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItemInput"
    PurchaseUpdate:
      type: object
      properties:
        invoice_no:
          type: string
        tax:
          type: number
        amount_paid:
          type: number
        bill_image_url:
          type: string
        description:
          type: string
        entry_date:
          type: string
          format: date
    PurchaseListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Purchase"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PurchaseItemInput:
      type: object
      required: [product_id, qty, unit_cost]
      properties:
        product_id:
          type: integer
        qty:
          type: number
        unit_cost:
          type: number
    PurchaseItem:
      type: object
      required: [id, product_id, product_name, qty, unit_cost]
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        qty:
          type: number
        unit_cost:
          type: number
    # ── Sales Order ───────────────────────────────────────────────────────────
...