# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * KhataPro CRM API — digital ledger platform for small businesses
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputUnit } from './productInputUnit';

export interface ProductInput {
  business_id: number;
  name: string;
  barcode?: string;
  sku?: string;
  category?: string;
  brand?: string;
  unit: ProductInputUnit;
  hsn_code?: string;
  gst_rate?: number;
  cost_price?: number;
  selling_price: number;
  stock_qty?: number;
  low_stock_alert?: number;
  image?: string;
}